postgres_to_es/requirements.txt
psycopg==3.1.18
python-dotenv==1.0.1

python-dateutil==2.9.0.post0
six==1.16.0